Changed registration version value to integer from unsigned long. Modified when the connection successful response is sent out from automatically to when the new connection asks for it.

+++ ChronoKeep/Constants/Network.cs
@@ -14,6 +14,6 @@ namespace Chronokeep.Constants
         public const int    CHRONOKEEP_ZCONF_PORT          = 4488;
 
         public const string CHRONOKEEP_REGISTRATION_TYPE = "CHRONOKEEP_WINDOWS";
-        public const ulong  CHRONOKEEP_REGISTRATION_VERS = 1;
+        public const int    CHRONOKEEP_REGISTRATION_VERS = 1;
     }
 }
